requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.3
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-filter==24.1
django-js-asset==2.2.0
django-storages==1.14.2
docutils==0.20.1
pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
whitenoise==6.6.0
